chore: fix version demoing-storybook
