+++ package.json
@@ -37,7 +37,8 @@
     "start": "npm run storybook",
     "storybook": "start-storybook -p 9001",
     "storybook:build": "build-storybook",
-    "test": "karma start --coverage",
+    "test": "karma start --coverage && npm run test:node",
+    "test:node": "mocha packages/convert-lion-modules/test/*.test.node.js",
     "test:watch": "karma start --auto-watch=true --single-run=false",
     "test:compatibility": "karma start --compatibility all --coverage",
     "test:compatibility:watch": "karma start --compatibility all --auto-watch=true --single-run=false",
