+++ packages/convert-lion-modules/index.js
@@ -1,0 +1,11 @@
+const { convertLionModules } = require('./src/convertLionModules.js');
+const { convertModule } = require('./src/convertModule.js');
+const {
+  createConvertLionModulesMiddleware,
+} = require('./src/createConvertLionModulesMiddleware.js');
+
+module.exports = {
+  convertModule,
+  convertLionModules,
+  createConvertLionModulesMiddleware,
+};
