+++ packages/convert-lion-modules/test/convertModule.test.node.js
@@ -1,8 +1,8 @@
 const { expect } = require('chai');
 
-const { replaceComponent } = require('../src/replaceComponent.js');
+const { convertModule } = require('../src/convertModule.js');
 
-describe('replaceComponent', () => {
+describe('convertModule', () => {
   const defaultConfig = {
     inTagName: 'lion-component',
     outTagName: 'ing-component',
@@ -16,25 +16,25 @@ describe('replaceComponent', () => {
       ...defaultConfig,
       getClassImportPath: () => '../../forms.js',
     };
-    expect(replaceComponent(input, config)).to.equal(output);
+    expect(convertModule(input, config)).to.equal(output);
   });
 
   it('2. replaces local src class imports', async () => {
     const input = `import { LionComponent } from '../src/LionComponent.js';`;
     const output = `import { IngComponent } from '../../../../packages/component/src/IngComponent.js';`;
-    expect(replaceComponent(input, defaultConfig)).to.equal(output);
+    expect(convertModule(input, defaultConfig)).to.equal(output);
   });
 
   it('3. replaces local tag imports', async () => {
     const input = `import '../lion-component.js';`;
     const output = `import '../../../../packages/component/ing-component.js';`;
-    expect(replaceComponent(input, defaultConfig)).to.equal(output);
+    expect(convertModule(input, defaultConfig)).to.equal(output);
   });
 
   it('4. replaces local index.js class imports', async () => {
     const input = `import { LionComponent, foo } from '../index.js';`;
     const output = `import { IngComponent, foo } from '../../../../packages/component/index.js';`;
-    expect(replaceComponent(input, defaultConfig)).to.equal(output);
+    expect(convertModule(input, defaultConfig)).to.equal(output);
   });
 
   it('5. replaces `@lion` imports', async () => {
@@ -46,7 +46,7 @@ describe('replaceComponent', () => {
       import { IngComponent, foo } from '../../../../packages/component/index.js';
       import '../../../../packages/component/ing-component.js';
     `;
-    expect(replaceComponent(input, defaultConfig)).to.equal(output);
+    expect(convertModule(input, defaultConfig)).to.equal(output);
   });
 
   it('6. replaces all remaining tag occurrences', async () => {
@@ -66,6 +66,6 @@ describe('replaceComponent', () => {
         <ing-component></ing-component>
       </ing-component>
     `;
-    expect(replaceComponent(input, defaultConfig)).to.equal(output);
+    expect(convertModule(input, defaultConfig)).to.equal(output);
   });
 });
