--- packages/replace-lion-features/demo/demo-calendar.js
@@ -1,18 +1,0 @@
-import { css } from '@lion/core';
-import { LionCalendar } from '@lion/calendar';
-
-class DemoCalendar extends LionCalendar {
-  static get styles() {
-    return [
-      super.styles,
-      css`
-        :host {
-          border-color: green !important;
-          color: red;
-        }
-      `,
-    ];
-  }
-}
-
-customElements.define('demo-calendar', DemoCalendar);
